build config to packages
