--- WorldOfPayback/Services/AppConfiguration/EnvironmentConfiguration.swift
@@ -1,14 +1,0 @@
-//
-//  EnvironmentConfiguration.swift
-//  WorldOfPayback
-//
-//  Created by Tsimur Asayonak on 12/20/23.
-//
-
-import Foundation
-
-enum EnvironmentConfiguration: String {
-    case development = "Development"
-    case production = "Production"
-    case staging = "Staging"
-}
